repofs: use underlying fs.download to download files
